init: add necessary dependencies
