+++ package.json
@@ -2,20 +2,21 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
+    "main": "src/index.ts",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start": "node src/index.ts",
+        "start:dev": "nodemon --exec ts-node src/index.ts"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "nodemon": "^2.0.16",
+        "typescript": "^5.6.3"
     },
     "dependencies": {
+        "dotenv": "^16.4.5",
+        "ts-node": "^10.9.2",
         "ws": "^8.8.0"
     }
 }
